refactor: reduce runtime event complexity
